requirements.txt
# requirements.txt
# Dependências do projeto Bot WhatsApp
# Compatível com Python 3.10.11

# === Framework Web ===
fastapi==0.104.1
uvicorn[standard]==0.24.0

# === Modelos de Dados ===
pydantic==2.5.0

# === LangChain (corrigido para versões compatíveis) ===
langchain-core==0.1.0
langchain-community==0.0.10

# === Cliente HTTP ===
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# === Manipulação de Arquivos ===
aiofiles==23.2.0
python-multipart==0.0.6

# === Processamento de Datas ===
python-dateutil==2.8.2

# === Configuração ===
python-dotenv==1.0.0

# === Logs Estruturados ===
structlog==23.2.0

# === QR Code para WhatsApp ===
qrcode[pil]==7.4.2

# === Processamento de Imagem (opcional) ===
Pillow==10.1.0

# === Utilitários ===
pathlib2==2.3.7; python_version < "3.4"

# === Desenvolvimento e Teste ===
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# === Monitoramento (opcional) ===
psutil==5.9.6